FRONT END/requirements.txt
python==3.10.8


Package                          Version
-------------------------------- ------------

numpy==1.23.5
pandas==2.2.3
scikit-learn==1.3.2
matplotlib==3.8.4
Flask==3.0.3
mysql-connector-python==9.1.0
lime==0.2.0.1
